+++ src/main/java/appeng/tile/misc/TileInterface.java
@@ -280,6 +280,11 @@ public class TileInterface extends AENetworkInvTile implements IGridTickable, II
     }
 
     @Override
+    public boolean refusedAsBusy() {
+        return this.duality.refusedAsBusy();
+    }
+
+    @Override
     public void provideCrafting(final ICraftingProviderHelper craftingTracker) {
         this.duality.provideCrafting(craftingTracker);
     }